refactor(rpc): move accounts handlers to routes/accounts.rs (backlog #12 phase 2e)
Nine address-indexed handlers come out of mod.rs into their own file:
balance, nonce, wallet info, tx list, richlist, address history, address
info, Merkle proof, and the per-height state root. sentrix_trie::address
imports move with them. Routes and responses are unchanged.

Two slices left after this: transactions and epoch.
